requirements.txt
Pillow==5.0.0
requests==2.18.4
tweepy==3.5.0
